requirements.txt
blinker==1.9.0
build==1.3.0
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
Flask==3.1.2
google-api-core==2.25.1
google-api-python-client==2.183.0
google-auth==2.41.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
httplib2==0.31.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
oauthlib==3.3.1
packaging==25.0
proto-plus==1.26.1
protobuf==6.32.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.5
pyproject_hooks==1.2.0
python-dotenv==1.1.1
redis==6.4.0
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
spotipy==2.25.1
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.1.3
